# openapi/v1.yaml
openapi: 3.1.0
info:
  title: Reverb HTTP API
  version: "1.0.0"
  summary: Semantic response cache with knowledge-aware invalidation.
  description: |
    Reverb exposes a two-tier semantic response cache (exact + semantic) over a
    versioned `/v1/` HTTP REST surface. This document is the authoritative
    contract for that surface; the Go HTTP handlers in `pkg/server/http.go`
    implement it, and the drift-check test in `pkg/server/openapi_drift_test.go`
    asserts the two stay aligned.

    ## Authentication

    When the standalone server is started with an `auth.Authenticator`, every
    endpoint except `/healthz`, `/readyz`, and `/metrics` requires a
    `Authorization: Bearer <token>` header. Missing or invalid tokens return
    `401 Unauthorized`.

    ## Rate limiting

    When a per-tenant rate limiter is configured, requests over the limit are
    rejected with `429 Too Many Requests` and a `Retry-After` header carrying
    the integer seconds the client should wait before retrying. `/healthz`,
    `/readyz`, and `/metrics` are never throttled.

    ## Versioning

    Backwards-incompatible changes ship under a new path prefix (`/v2/`). Within
    `/v1/` the contract is additive: new optional fields may appear, existing
    fields will not be removed or have their types changed.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Reverb maintainers
    url: https://github.com/nobelk/reverb
servers:
...
security:
  - bearerAuth: []
  - {}
...
paths:
  /v1/lookup:
    post:
      tags: [cache]
      summary: Look up a cached response for a prompt.
      description: |
        Performs an exact-match check first (SHA-256 of the normalized prompt),
        then a semantic-similarity search if the exact tier misses. Returns
        `hit=false` when both tiers miss; the caller is then expected to invoke
        the upstream model and store the response via `POST /v1/store`.
      operationId: lookup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LookupRequest"
      responses:
        "200":
          description: Lookup completed (hit or miss; check the `hit` field).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LookupResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/lookup-stream:
    post:
      tags: [cache]
      summary: Stream a cached response as Server-Sent Events.
      description: |
        On a cache hit, replays the entry as `text/event-stream`. Each chunk
        becomes a single `data:` line; the stream terminates with `data: [DONE]`
        per the OpenAI SSE convention. Entries stored without `chunks` are
        emitted as a single chunk carrying the full response text. A 15s
        comment-line keepalive is interleaved so intermediate proxies don't
        time out long replays. On a miss, returns `404` with the standard
        error envelope; the response body is JSON, not SSE.
      operationId: lookupStream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LookupRequest"
      responses:
        "200":
          description: Cache hit; body is an SSE stream of `ResponseChunk` objects terminated by `[DONE]`.
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/store:
    post:
      tags: [cache]
      summary: Store a response in the cache.
      description: |
        Persists a prompt/response pair plus optional source lineage. The entry
        is added to both the exact-match tier and the semantic tier (an
        embedding is computed via the configured provider).
      operationId: store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreRequest"
      responses:
        "201":
          description: Entry stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invalidate:
    post:
      tags: [lineage]
      summary: Invalidate every entry derived from a source.
      description: |
        Removes all cache entries whose lineage references `source_id`. This is
        the entry point used by CDC listeners when a source document changes
        upstream.
      operationId: invalidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvalidateRequest"
      responses:
        "200":
          description: Invalidation completed (count may be zero).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/entries/{id}:
    delete:
      tags: [cache]
      summary: Delete a single cache entry.
      description: |
        Removes the entry from both the persistence store and the vector index.
        Under tenant-scoped auth, returns `404` if the entry does not belong to
        the requesting tenant (without revealing which case it is).
      operationId: deleteEntry
      parameters:
        - name: id
          in: path
          required: true
          description: The entry ID returned by `POST /v1/store`.
          schema:
            type: string
      responses:
        "204":
          description: Entry deleted (no body).
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/stats:
    get:
      tags: [ops]
      summary: Cache statistics.
      description: |
        Returns counters and the namespace list. Under tenant-scoped auth, the
        response is filtered to namespaces owned by the requesting tenant and
        `total_entries` is recomputed against that subset.
      operationId: getStats
      responses:
        "200":
          description: Stats snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Opaque bearer token issued by the operator. Reverb does not issue tokens
        itself; the server is configured with an `auth.Authenticator` that
        validates tokens out-of-band.
  headers:
    RetryAfter:
      description: Integer seconds the client should wait before retrying.
      schema:
        type: integer
        minimum: 1
  responses:
    BadRequest:
      description: Malformed request (invalid JSON, missing required field, bad content_hash).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource does not exist or is not visible to the caller.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Per-tenant rate limit exceeded.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Unhandled server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    LookupRequest:
      type: object
      required: [namespace, prompt]
      properties:
        namespace:
          type: string
          description: Logical partition. Required.
          example: support-bot
        prompt:
          type: string
          description: Raw prompt text. Required. Normalized server-side.
          example: How do I reset my password?
        model_id:
          type: string
          description: Model identifier. When `scope_by_model` is true, restricts semantic matches to the same model.
          example: gpt-4o
    LookupResponse:
      type: object
      required: [hit]
      properties:
        hit:
          type: boolean
          description: True iff either tier returned a match.
        tier:
          type: string
          description: Which tier produced the hit. Empty on miss.
          enum: [exact, semantic]
        similarity:
          type: number
          format: float
          description: Cosine similarity for semantic hits. Omitted on exact hits and misses.
          minimum: 0
          maximum: 1
        entry:
          $ref: "#/components/schemas/CacheEntry"
    StoreRequest:
      type: object
      required: [namespace, prompt]
      properties:
        namespace:
          type: string
          description: Logical partition. Required.
        prompt:
          type: string
          description: Raw prompt text. Required.
        model_id:
          type: string
          description: Model identifier (recommended).
        response:
          type: string
          description: |
            The model's response text. Either `response` or `chunks` is required.
            When both are present, `chunks` wins and `response` is overridden by
            the concatenation of every chunk's `delta`.
        chunks:
          type: array
          description: |
            Streamed-delta form of the response. Either `response` or `chunks`
            is required. When present, the entry is replayable via
            `POST /v1/lookup-stream`.
          items:
            $ref: "#/components/schemas/ResponseChunk"
        response_meta:
          type: object
          description: Free-form metadata attached to the entry.
          additionalProperties:
            type: string
        sources:
          type: array
          description: Source-document lineage. Drives invalidation when sources change.
          items:
            $ref: "#/components/schemas/SourceRef"
        ttl_seconds:
          type: integer
          minimum: 0
          description: Override the default TTL. Zero or omitted uses the server default.
    ResponseChunk:
      type: object
      required: [delta]
      properties:
        delta:
          type: string
          description: The text contribution of this chunk.
        finish_reason:
          type: string
          description: Set on the terminal chunk only. Mirrors OpenAI semantics ("stop", "length", "content_filter", "tool_calls", ...).
    StoreResponse:
      type: object
      required: [id, created_at]
      properties:
        id:
          type: string
          description: Server-assigned entry ID (UUID).
        created_at:
          type: string
          format: date-time
    InvalidateRequest:
      type: object
      required: [source_id]
      properties:
        source_id:
          type: string
          description: The source document whose dependents should be invalidated.
    InvalidateResponse:
      type: object
      required: [invalidated_count]
      properties:
        invalidated_count:
          type: integer
          minimum: 0
          description: Number of cache entries removed.
    StatsResponse:
      type: object
      required:
        - total_entries
        - namespaces
        - exact_hits_total
        - semantic_hits_total
        - misses_total
        - invalidations_total
        - hit_rate
      properties:
        total_entries:
          type: integer
          format: int64
          minimum: 0
        namespaces:
          type: array
          items:
            type: string
        exact_hits_total:
          type: integer
          format: int64
          minimum: 0
        semantic_hits_total:
          type: integer
          format: int64
          minimum: 0
        misses_total:
          type: integer
          format: int64
          minimum: 0
        invalidations_total:
          type: integer
          format: int64
          minimum: 0
        hit_rate:
          type: number
          format: double
          minimum: 0
          maximum: 1
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message. Stable enough to log; not stable enough to programmatically match on.